# Hoist variable in lu out of loop